(SERVER-1406) Add 'medium' catalog
This commit adds a new role called `by_size::medium`.  The catalog
generated by this role contains many more `package`, `user`, and
other resources in an effort to get us up to something a bit more
real-world.  The contents of the catalog were loosely based on
a real catalog that Jeremy compiled from ops, so the various counts
of different resource types are in the general ballpark of the
ones that we saw in that catalog.  For more details see the
Jira ticket.